requirements.txt
pytest==7.4.3
python-dotenv==1.0.0
flake8==6.1.0
black==23.11.0
ipython==8.17.2
flask==3.0.0
flask-cors==4.0.0
chromadb==0.4.18
numpy<2.0.0
PyJWT==2.8.0
flask-limiter==3.5.0
python-json-logger==2.0.7
bcrypt==4.1.2
flask-login==0.6.3